backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
pytest==7.4.3
black==23.11.0

